+++ data/animated_player_trainers.lua
@@ -19,4 +19,8 @@ return {
   ash  = five("ashplayer.png"),
   gary = five("garyplayer.png"),
   red  = five("redplayer.png"),
+  ash_front   = five("ashfrontplayer.png"),
+  brock_front = five("brockfrontplayer.png"),
+  bulma_front = five("bulmafrontplayer.png"),
+  gary_front  = five("garyfrontplayer.png"),
 }